requirements.txt
flask
requests
PyPDF2
google-genai
pycryptodome
gunicorn